Fix issues with roles
