+++ app/javascript/dashboard/components/layout/sidebarComponents/SecondaryNavItem.vue
@@ -1,93 +1,128 @@
 <template>
-  <router-link
-    v-slot="{ href, isActive, navigate }"
-    :to="to"
-    custom
-    active-class="active"
-  >
-    <li :class="{ active: isActive }">
-      <a
-        :href="href"
-        class="button clear menu-item text-truncate"
-        :class="{ 'is-active': isActive, 'text-truncate': shouldTruncate }"
-        @click="navigate"
+  <li :class="computedClass" class="sidebar-item">
+    <span v-if="hasSubMenu" class="secondary-menu--title fs-small">
+      {{ $t(`SIDEBAR.${menuItem.label}`) }}
+    </span>
+    <router-link
+      v-else
+      class="secondary-menu--title secondary-menu--link fs-small"
+      :to="menuItem.toState"
+    >
+      <fluent-icon :icon="menuItem.icon" class="secondary-menu--icon" />
+      {{ $t(`SIDEBAR.${menuItem.label}`) }}
+    </router-link>
+
+    <ul v-if="hasSubMenu" class="nested vertical menu">
+      <secondary-child-nav-item
+        v-for="child in menuItem.children"
+        :key="child.id"
+        :to="child.toState"
+        :label="child.label"
+        :label-color="child.color"
+        :should-truncate="child.truncateLabel"
+        :icon="computedInboxClass(child)"
+      />
+      <router-link
+        v-if="showItem(menuItem)"
+        v-slot="{ href, isActive, navigate }"
+        :to="menuItem.toState"
+        custom
       >
-        <span v-if="icon" class="badge--icon">
-          <fluent-icon class="inbox-icon" :icon="icon" size="10" />
-        </span>
-        <span
-          v-if="labelColor"
-          class="badge--label"
-          :style="{ backgroundColor: labelColor }"
-        />
-        <span
-          :title="menuTitle"
-          class="menu-label button__content"
-          :class="{ 'text-truncate': shouldTruncate }"
-        >
-          {{ label }}
-        </span>
-        <span v-if="count" class="badge" :class="{ secondary: !isActive }">
-          {{ count }}
-        </span>
-      </a>
-    </li>
-  </router-link>
+        <li>
+          <a
+            :href="href"
+            class="button small clear menu-item--new secondary"
+            :class="{ 'is-active': isActive }"
+            @click="e => newLinkClick(e, navigate)"
+          >
+            <fluent-icon icon="add" />
+            <span class="button__content">
+              {{ $t(`SIDEBAR.${menuItem.newLinkTag}`) }}
+            </span>
+          </a>
+        </li>
+      </router-link>
+    </ul>
+  </li>
 </template>
+
 <script>
+import { mapGetters } from 'vuex';
+
+import adminMixin from '../../../mixins/isAdmin';
+import { getInboxClassByType } from 'dashboard/helper/inbox';
+
+import SecondaryChildNavItem from './SecondaryChildNavItem';
+
 export default {
+  components: { SecondaryChildNavItem },
+  mixins: [adminMixin],
   props: {
-    to: {
-      type: String,
-      default: '',
-    },
-    label: {
-      type: String,
-      default: '',
-    },
-    labelColor: {
-      type: String,
-      default: '',
-    },
-    shouldTruncate: {
-      type: Boolean,
-      default: false,
-    },
-    icon: {
-      type: String,
-      default: '',
-    },
-    count: {
-      type: String,
-      default: '',
+    menuItem: {
+      type: Object,
+      default: () => ({}),
     },
   },
   computed: {
-    showIcon() {
-      return { 'text-truncate': this.shouldTruncate };
+    ...mapGetters({ activeInbox: 'getSelectedInbox' }),
+    hasSubMenu() {
+      return !!this.menuItem.children;
     },
-    menuTitle() {
-      return this.shouldTruncate ? this.label : '';
+    computedClass() {
+      // If active Inbox is present
+      // donot highlight conversations
+      if (this.activeInbox) return ' ';
+
+      if (
+        this.$store.state.route.name === 'inbox_conversation' &&
+        this.menuItem.toStateName === 'home'
+      ) {
+        return 'active';
+      }
+      return ' ';
+    },
+  },
+  methods: {
+    computedInboxClass(child) {
+      const { type, phoneNumber } = child;
+      if (!type) return '';
+      const classByType = getInboxClassByType(type, phoneNumber);
+      return classByType;
+    },
+    newLinkClick(e, navigate) {
+      if (this.menuItem.newLinkRouteName) {
+        navigate(e);
+      } else if (this.menuItem.showModalForNewItem) {
+        if (this.menuItem.modalName === 'AddLabel') {
+          e.preventDefault();
+          this.$emit('add-label');
+        }
+      }
+    },
+    showItem(item) {
+      return this.isAdmin && item.newLink !== undefined;
     },
   },
 };
 </script>
 <style lang="scss" scoped>
-$badge-size: var(--space-slab);
-
-.button {
-  margin: var(--space-small) 0;
+.sidebar-item {
+  margin: var(--space-smaller) 0 0;
 }
 
-.menu-item {
-  display: inline-flex;
+.secondary-menu--title {
   color: var(--s-600);
+  display: flex;
+  font-weight: var(--font-weight-bold);
+  line-height: var(--space-two);
+  margin: var(--space-small) 0;
+  padding: 0 var(--space-small);
+}
+
+.secondary-menu--link {
   font-weight: var(--font-weight-medium);
-  width: 100%;
-  height: var(--space-medium);
-  padding: var(--space-smaller) var(--space-smaller);
-  margin: var(--space-smaller) 0;
-  text-align: left;
+  margin: 0;
+  padding: var(--space-small);
 
   &:hover {
     background: var(--s-25);
@@ -98,41 +133,49 @@ $badge-size: var(--space-slab);
     border-color: var(--w-300);
   }
 
-  &.is-active {
+  &.router-link-exact-active {
     background: var(--w-25);
     color: var(--w-500);
     border-color: var(--w-25);
   }
 }
 
-.menu-label {
-  flex-grow: 1;
-  line-height: var(--space-two);
+.secondary-menu--icon {
+  margin-right: var(--space-small);
+  min-width: var(--space-normal);
+}
+
+.sub-menu-link {
+  color: var(--s-600);
+}
+
+.wrap {
+  display: flex;
+  align-items: center;
+}
+
+.label-color--display {
+  border-radius: var(--space-smaller);
+  height: var(--space-normal);
+  margin-right: var(--space-small);
+  min-width: var(--space-normal);
+  width: var(--space-normal);
 }
 
 .inbox-icon {
-  font-size: var(--font-size-nano);
+  position: relative;
+  top: -1px;
 }
 
-.badge--label,
-.badge--icon {
+.sidebar-item .button.menu-item--new {
   display: inline-flex;
-  min-width: $badge-size;
-  height: $badge-size;
-  border-radius: var(--border-radius-small);
-  margin-right: var(--space-smaller);
-  background: var(--s-100);
-}
+  height: var(--space-medium);
+  margin: var(--space-smaller) 0;
+  padding: var(--space-smaller);
+  color: var(--s-500);
 
-.badge--icon {
-  align-items: center;
-  justify-content: center;
-}
-
-.badge.secondary {
-  min-width: unset;
-  background: var(--s-75);
-  color: var(--s-600);
-  font-weight: var(--font-weight-bold);
+  &:hover {
+    color: var(--w-500);
+  }
 }
 </style>
